Add tests, generalize the margin rule to radial chart
